requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2018.8.24
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==2.6.1
entrypoints==0.3
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
oauthlib==3.2.2
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4.3
python-dotenv==0.21.1
pytz==2023.4
pyxdg==0.25
requests==2.31.0
requests-oauthlib==1.3.1
SecretStorage==2.3.1
six==1.12.0
soupsieve==2.4.1
ssh-import-id==5.7
tweepy==4.14.0
urllib3==1.24.1